Scripts for app attach are now signed and work. Additionally, configuration is in JSON file, and the naming has been simplified (stage.ps1 etc.)

+++ artifacts/templates/AppAttachDeregister.ps1
@@ -1,13 +1,13 @@
-#MSIX app attach deregistration sample
+$configFilePath = (Get-ChildItem $PSScriptRoot -Filter *.json | Select-Object -First 1).FullName;
+if ($null -eq $configFilePath)
+{
+    throw "Missing JSON config!";
+}
 
-#region variables
+$configFile = Get-Content $configFilePath -Raw | ConvertFrom-Json;
 
-$packageName = "<package name>"
-
-#endregion
-
-#region deregister
-
-Remove-AppxPackage -PreserveRoamableApplicationData $packageName
-
-#endregion
+foreach ($package in $configFile)
+{
+    $packageName = $package.packageName;
+    Remove-AppxPackage -PreserveRoamableApplicationData $packageName
+}
